fix projects index
